+++ apps/laboratory/tests/shared/utils/project.ts
@@ -1,5 +1,7 @@
 import { devices } from '@playwright/test'
 import { getAvailableDevices } from './device'
+import { getValue } from './config'
+import { getLocalBravePath, BRAVE_LINUX_PATH } from '../constants/browsers'
 
 const availableDevices = getAvailableDevices()
 
@@ -9,16 +11,35 @@ const PERMUTATIONS = availableDevices.flatMap(device =>
   LIBRARIES.map(library => ({ device, library }))
 )
 
+interface UseOptions {
+  launchOptions: {
+    executablePath: string
+  }
+}
+
 interface CustomProperties {
   testIgnore?: string
   testMatch?: string
+  useOptions?: UseOptions
 }
 
 export type CustomProjectProperties = {
   [T in string]: CustomProperties
 }
 
+const braveOptions: UseOptions = {
+  launchOptions: {
+    executablePath: getValue(BRAVE_LINUX_PATH, getLocalBravePath())
+  }
+}
+
 const customProjectProperties: CustomProjectProperties = {
+  'Desktop Brave/wagmi': {
+    useOptions: braveOptions
+  },
+  'Desktop Brave/ethers': {
+    useOptions: braveOptions
+  },
   'Desktop Chrome/wagmi': {
     testIgnore: 'email.spec.ts'
   },
@@ -37,13 +58,17 @@ export interface Permutation {
 
 export function getProjects() {
   return PERMUTATIONS.map(({ device, library }) => {
+    const deviceName = device === 'Desktop Brave' ? 'Desktop Chrome' : device
     let project = {
       name: `${device}/${library}`,
-      use: { ...devices[device], library }
+      use: { ...devices[deviceName], library }
     }
     const props = customProjectProperties[project.name]
     if (props) {
       project = { ...project, ...props }
+      if (props.useOptions) {
+        project.use = { ...project.use, ...props.useOptions }
+      }
     }
 
     return project