+++ packages/siwe/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @web3modal/siwe
 
+## 4.0.11
+
+### Patch Changes
+
+- Analytics connection event improvements. Unsupported chain flag. Siwe package refactor. RPC improvements. UI improvements'
+
+- Updated dependencies []:
+  - @web3modal/core@4.0.11
+  - @web3modal/scaffold-utils@4.0.11
+
 ## 4.0.10
 
 ### Patch Changes